refactor: commonStyles 위치 변경
