+++ src/components/common/commonStyles.ts
